tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-two-iframes-manual.html

<!DOCTYPE html>
<title>Element#requestFullscreen() on two elements in different iframes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<iframe id="a" allowfullscreen></iframe>
<iframe id="b" allowfullscreen></iframe>
<script>
async_test(t => {
  // Request fullscreen on the body elements of both iframes, but in reverse
  // tree order.
  const a = document.getElementById('a');
  const b = document.getElementById('b');

  // Expect two fullscreenchange events, with document.fullscreenElement
  // changing in the same order as the requests. (Events should also fire on the
  // iframes' documents, but this is not covered by this test.)
  const order = [];
  document.onfullscreenchange = t.step_func(() => {
    assert_in_array(document.fullscreenElement, [a, b]);
    order.push(document.fullscreenElement.id);
    if (order.length == 2) {
      // Since fullscreenchange event occurs at animation frame timing we might
      // have not seen the transition from null -> 'b' but just see the
      // resulting 'a' transition twice.
      assert_true(order[0] == 'a' || order[0] == 'b', 'first id seen is a or b');
      assert_true(order[1] == 'a', 'second id seen is b');
      t.done();
    }
  });
  document.onfullscreenerror = t.unreached_func('fullscreenerror event');

  trusted_click(t, () => {
    b.contentDocument.body.requestFullscreen();
    a.contentDocument.body.requestFullscreen();
  }, document.body);
});
</script>
